+++ packages/theme-check-browser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @shopify/theme-check-browser
 
+## 2.6.0
+
+### Patch Changes
+
+- Updated dependencies [ff78229]
+  - @shopify/theme-check-common@2.6.0
+
 ## 2.5.1
 
 ### Patch Changes